接口测试

+++ http-download-server/src/main/java/com/example/httpdownloadserver/control/SseController.java
@@ -19,7 +19,7 @@ public class SseController {
 
     @GetMapping("/download/sse")
     public void sendSseMessage(String taskId) throws IOException {
-        TaskDO taskDO = taskDAO.selectById(Long.parseLong(taskId));
+        TaskDO taskDO = taskDAO.selectById(Integer.parseInt(taskId));
         Task task = taskDO.toModel();
         downloadService.download(task, task.getDownloadThread());
     }
