+++ http-download-server/src/main/java/com/example/httpdownloadserver/dao/TaskDAO.java
@@ -9,10 +9,10 @@ public interface TaskDAO {
     int insert(TaskDO taskDO);
 
     //根据任务id查询任务
-    TaskDO selectById(Long id);
+    TaskDO selectById(Integer id);
 
     //根据任务id删除任务
-    int deleteById(Long id);
+    int deleteById(Integer id);
 
     //查询所有任务
     List<TaskDO> selectAll();
@@ -20,12 +20,12 @@ public interface TaskDAO {
     //根据任务状态查询任务
     List<TaskDO> selectByStatus(String status);
     //根据所选列表删除文件
-    int deleteByIds(List<Long> ids);
+    int deleteByIds(List<Integer> ids);
     //根据任务id更新任务
-    int updateThreadById(Long id,int downloadThread);
+    int updateThreadById(Integer id,int downloadThread);
     //获得线程数
-    int getThreadById(Long id);
+    int getThreadById(Integer id);
     //根据任务id更新任务
-    int updateById(Long id);
+    int updateById(Integer id);
 
 }
